Feature google auth (#16)
* Google auth
* Delete useless dependencies
* Change id to string
* Create google authentication
* Set google redirect url to properties
* Create constant fields

+++ src/main/java/com/project/devblog/model/CommentEntity.java
@@ -19,11 +19,11 @@ import java.time.LocalDateTime;
 @FieldDefaults(level = AccessLevel.PRIVATE)
 @Entity
 @Table(name = "comments")
-public class CommentEntity extends AuditableBaseEntity<Integer> {
+public class CommentEntity extends AuditableBaseEntity<Long> {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    Integer id;
+    Long id;
 
     String message;
     Boolean enabled;
